+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### CI/CD
 
 ### Dependencies
-* Bump cicirello/pyaction from 3.13.5-gh-2.75.1 to 3.13.5-gh-2.76.2
+* Bump cicirello/pyaction from 3.13.5-gh-2.75.1 to 3.13.6-gh-2.76.2
 
 
 ## [1.10.2] - 2025-07-15